Remove unwanted status field from Event and refactor tests
